// clients/admin-ui/src/types/api/models/Page_DatastoreStagedResourceTreeAPIResponse_.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { DatastoreStagedResourceTreeAPIResponse } from './DatastoreStagedResourceTreeAPIResponse';

export type Page_DatastoreStagedResourceTreeAPIResponse_ = {
  items: Array<DatastoreStagedResourceTreeAPIResponse>;
  total: (number | null);
  page: (number | null);
  size: (number | null);
  pages?: (number | null);
};